set wallet icon to wallet if not connected
