+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "hub-hoprnet-org",
-  "version": "1.7.1",
+  "version": "1.7.2",
   "private": true,
   "dependencies": {
     "@emotion/react": "^11.13.3",
